Safari-609.1.11.1/LayoutTests/http/tests/security/xssAuditor/malformed-xss-protection-header-2-expected.txt

CONSOLE MESSAGE: line 1: Error parsing header X-XSS-Protection: red: expected 0 or 1 at character position 0. The default protections will be applied.
CONSOLE MESSAGE: line 4: The XSS Auditor refused to execute a script in 'http://127.0.0.1:8000/security/xssAuditor/resources/echo-intertag.pl?test=/security/xssAuditor/malformed-xss-protection-header-2.html&notifyDone=1&malformed-header=2&q=%3Cscript%3Ealert(String.fromCharCode(0x58,0x53,0x53))%3C/script%3E%3Cp%3EIf%20you%20see%20this%20message%20and%20no%20JavaScript%20alert()%20then%20the%20test%20PASSED.%3C/p%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
This tests that the X-XSS-Protection header is not ignored when the first character is not 0 or 1, and that we issue an error.



--------
Frame: 'frame'
--------
If you see this message and no JavaScript alert() then the test PASSED.
